aws_serverless_data_engineering/requirements.txt
aws-cdk-lib==2.50.0
constructs>=10.0.0,<11.0.0
requests==2.27.1
pandas==1.4.2
pyarrow==8.0.0
fsspec==2022.5.0
s3fs<=0.4